Minor code cleanup.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/branches/irrlicht@3462 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/audio/sfx_manager.cpp
@@ -169,7 +169,7 @@ bool loadVorbisBuffer(const char *name, ALuint buffer)
         {
             info = ov_info(&oggFile, -1);
             
-            long len = ov_pcm_total(&oggFile, -1) * info->channels * 2;    // always 16 bit data
+            long len = (long)ov_pcm_total(&oggFile, -1) * info->channels * 2;    // always 16 bit data
             
             char *data = (char *) malloc(len);
